backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.6.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]
passlib[argon2]
python-multipart
aiofiles
groq
python-dotenv
httpx
email-validator
